chore: refine decoupled func name and docs (#28)
