+++ rlinf/scheduler/__init__.py
@@ -35,14 +35,18 @@ from .placement import (
 )
 from .worker import Worker, WorkerAddress, WorkerGroupFuncResult
 from .worker.routing import (
-    RouteEntry,
-    RoutePlan,
+    CommMapper,
     build_recv_plan,
     build_route_channel_key,
+    build_send_key,
     build_send_plan,
+    decoupled_build_recv_plan,
+    get_batch_size,
+    get_group_world_size,
     infer_batch_size,
     merge_batches,
     split_batch,
+    split_channel_message,
 )
 
 __all__ = [
@@ -65,11 +69,15 @@ __all__ = [
     "Worker",
     "WorkerAddress",
     "WorkerGroupFuncResult",
-    "RouteEntry",
-    "RoutePlan",
+    "CommMapper",
+    "split_channel_message",
     "build_send_plan",
+    "build_send_key",
     "build_recv_plan",
     "build_route_channel_key",
+    "decoupled_build_recv_plan",
+    "get_batch_size",
+    "get_group_world_size",
     "infer_batch_size",
     "split_batch",
     "merge_batches",
