+++ rlinf/scheduler/worker/__init__.py
@@ -13,27 +13,49 @@
 # limitations under the License.
 
 from .routing import (
+    CommMapper,
+    DecoupledRouteEntry,
     RouteEntry,
     RoutePlan,
+    _build_channel_message,
     build_recv_plan,
     build_route_channel_key,
+    build_send_key,
     build_send_plan,
+    decoupled_build_recv_plan,
+    decoupled_build_send_plan,
+    get_batch_size,
+    get_group_info,
+    get_group_world_size,
     infer_batch_size,
     merge_batches,
     split_batch,
+    split_channel_message,
+    validate_batch_size,
 )
 from .worker import Worker, WorkerAddress
 from .worker_group import WorkerGroup, WorkerGroupFunc, WorkerGroupFuncResult
 
 __all__ = [
+    "CommMapper",
+    "DecoupledRouteEntry",
     "RouteEntry",
     "RoutePlan",
+    "_build_channel_message",
+    "split_channel_message",
     "build_recv_plan",
     "build_route_channel_key",
+    "build_send_key",
     "build_send_plan",
+    "decoupled_build_recv_plan",
+    "decoupled_build_send_plan",
+    "get_batch_size",
+    "get_group_info",
+    "get_group_world_size",
     "infer_batch_size",
     "merge_batches",
     "split_batch",
+    "validate_batch_size",
     "Worker",
     "WorkerAddress",
     "WorkerGroup",
