+++ rlinf/workers/reward/reward_worker.py
@@ -376,7 +376,7 @@ class EmbodiedRewardWorker(Worker):
                 tag="train_reward_obs",
                 async_op=True,
                 batch_size=self.train_batch_size,
-                env_decoupled_mode=self.env_decoupled_mode,
+                decoupled_mode=self.env_decoupled_mode,
             ).async_wait()
             rewards = self.compute_image_rewards(observations=merged_data)
             if isinstance(rewards, torch.Tensor):
@@ -387,7 +387,7 @@ class EmbodiedRewardWorker(Worker):
                 data=rewards,
                 tag="train_reward_obs",
                 async_op=True,
-                env_decoupled_mode=self.env_decoupled_mode,
+                decoupled_mode=self.env_decoupled_mode,
             )
 
     async def stop(self):
